move quicklook-custom files inside desispec/quicklook
